Fixed ValueFormatter to handle unknown floatfield values
If the user specified an unknown floatfield value, say "fxed" instead
of "fixed", the code would rely on an uninitialized character which
would produce a garbage format string, which could lead to unknown
behaviour in libfmt. Instead, the code will now simply silently leave
the floatfield part of the format string empty.