Cast __PRETTY_FUNCTION__ to char * to avoid code analysis warnings
Code analysis software prefers all array accesses to be spans so that
indexes can be checked. Hence they warn if an array of char is
converted to a plain char pointer with which no such checking can be
done.

Unfortunately, __PRETTY_FUNCTION__ is defined to be an array, and hence
using it in Spine::Exception generates a lot of warnings. The macro
should be replaced with std::source_location, but it is not currently
available. Hence we use an explicit cast to silence the warning.

Further discussion: https://github.com/isocpp/CppCoreGuidelines/issues/765