Do not print unnecessary error information to the journal
If the stack trace is disabled, the Exception::printError will now
print nothing. Previously only the stack trace was omitted, but the
error message, URI and any parameters set to the exception such as
client IP were printed. The idea is that if the stack trace is disabled,
the error is most likely due to bad input parameters and hence
printing any information to the journal is unnecessary. The access
log will still mark the URI with an HTTP error status code.